Evm::execute general output
