+++ src/eth/executor/types/mod.rs
@@ -1,5 +1,5 @@
+mod account_changes;
 mod error;
-mod execution_account_changes;
 mod execution_changes;
 mod execution_metrics;
 mod execution_output;
@@ -7,9 +7,9 @@ mod execution_result;
 mod task;
 mod transaction_execution;
 
+pub use account_changes::AccountChanges;
+pub use account_changes::ChangeValue;
 pub use error::ExecutorError;
-pub use execution_account_changes::AccountChangeValue;
-pub use execution_account_changes::ExecutionAccountChanges;
 pub use execution_changes::AccountOriginalsReader;
 pub use execution_changes::Complete;
 pub use execution_changes::ExecutionChanges;
