+++ src/eth/storage/stratus_storage.rs
@@ -849,7 +849,7 @@ impl StratusStorage {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::eth::executor::ExecutionAccountChanges;
+    use crate::eth::executor::AccountChanges;
     use crate::eth::executor::ExecutionResult;
     use crate::eth::executor::TransactionExecutionInput;
     use crate::eth::executor::TransactionExecutionOutput;
@@ -914,18 +914,16 @@ mod tests {
 
         // Mine a block setting the account balance to 100. The eth_call captures this block.
         let mut changes1 = ExecutionChanges::default();
-        changes1.accounts.insert(
-            address,
-            ExecutionAccountChanges::from_changed(Account::new_with_balance(address, Wei::from(100u64))),
-        );
+        changes1
+            .accounts
+            .insert(address, AccountChanges::from_changed(Account::new_with_balance(address, Wei::from(100u64))));
         let call_block = mine_block(&storage, changes1);
 
         // A new block is mined while the call is in flight, changing the balance to 200.
         let mut changes2 = ExecutionChanges::default();
-        changes2.accounts.insert(
-            address,
-            ExecutionAccountChanges::from_changed(Account::new_with_balance(address, Wei::from(200u64))),
-        );
+        changes2
+            .accounts
+            .insert(address, AccountChanges::from_changed(Account::new_with_balance(address, Wei::from(200u64))));
         let latest = mine_block(&storage, changes2);
         assert_ne!(call_block, latest);
