+++ src/eth/types/mod.rs
@@ -62,8 +62,8 @@ pub use transaction::TransactionStage;
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::eth::executor::AccountChanges;
     use crate::eth::executor::EvmExecutionMetrics;
-    use crate::eth::executor::ExecutionAccountChanges;
     use crate::eth::executor::ExecutionResult;
     use crate::eth::rpc::BlockFilter;
     use crate::eth::rpc::LogFilter;
@@ -102,7 +102,7 @@ mod tests {
     gen_test_serde!(EcdsaRs);
     gen_test_serde!(EcdsaV);
     gen_test_serde!(EvmExecutionMetrics);
-    gen_test_serde!(ExecutionAccountChanges);
+    gen_test_serde!(AccountChanges);
     gen_test_serde!(ExecutionResult);
     gen_test_serde!(Gas);
     gen_test_serde!(Hash);
